+++ src/plinkliftover/__init__.py
@@ -6,3 +6,24 @@ try:
     __version__ = version(__name__)
 except PackageNotFoundError:  # pragma: no cover
     __version__ = "unknown"
+
+import typer
+from rich.console import Console
+
+console = Console()
+
+
+def version_callback(value: bool):
+    """Prints the version of the package."""
+    if value:
+        console.print(
+            f"[yellow]plinkliftover[/] version: [bold blue]{__version__}[/]"
+        )
+        raise typer.Exit()
+
+
+app = typer.Typer(
+    name="plinkliftover",
+    help="Converts genotype data stored in plink's PED+MAP format from one genome build to another, using liftOver",
+    add_completion=False,
+)