+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2021-02-02
+## [0.3.0] - 2022-02-17
+
+### Added
+
+- Added cli interface for map2bed and bed2map
+
+### Changed
+
+- Split off `map2bed` and `bed2map` into separate submodules
+
+## [0.2.0] - 2022-02-02
 
 ### Added
 
@@ -15,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed conditional requirement of `import_metadata` as a version of Python >=3.8 is required
 
-## [0.1.11] - 2021-02-02
+## [0.1.11] - 2022-02-02
 
 ### Changed